Fix errors in test. Refactor some code
