+++ java/src/jmri/jmrit/logixng/FemaleGenericExpressionSocket.java
@@ -56,43 +56,6 @@ public interface FemaleGenericExpressionSocket
     /**
      * Evaluate this expression.
      * <P>
-     * The return value of the evaluation is converted to a boolean if necessary.
-     * <P>
-     * The parameter isCompleted is used if the expression should be evaluated
-     * more than once. For example, the Count expression is not completed until
-     * its child expression has been true and false a number of times.
-     * 
-     * @return the result of the evaluation
-     * @throws jmri.JmriException when an exception occurs
-     */
-    public boolean evaluateBoolean() throws JmriException;
-    
-    /**
-     * Evaluate this expression.
-     * <P>
-     * The return value of the evaluation is converted to a double if necessary.
-     * 
-     * @return the result of the evaluation. The male socket that holds this
-     * expression throws an exception if this value is a Double.NaN or an
-     * infinite number.
-     * 
-     * @throws jmri.JmriException when an exception occurs
-     */
-    public double evaluateDouble() throws JmriException;
-    
-    /**
-     * Evaluate this expression.
-     * <P>
-     * The return value of the evaluation is converted to a String if necessary.
-     * 
-     * @return the result of the evaluation
-     * @throws jmri.JmriException when an exception occurs
-     */
-    public String evaluateString() throws JmriException;
-    
-    /**
-     * Evaluate this expression.
-     * <P>
      * This method validates the expression without doing any convertation of
      * the return value.
      * <P>
