+++ java/src/jmri/jmrit/logixng/expressions/StringFormula.java
@@ -98,8 +98,7 @@ public class StringFormula extends AbstractStringExpression implements FemaleSoc
             Base parent, FemaleSocketListener listener, String socketName) {
         
         return new DefaultFemaleGenericExpressionSocket(
-                FemaleGenericExpressionSocket.SocketType.GENERIC, parent, listener, socketName)
-                .getGenericSocket(parent);
+                FemaleGenericExpressionSocket.SocketType.GENERIC, parent, listener, socketName);
     }
 
     public final void setFormula(String formula) throws ParserException {
