+++ java/src/jmri/jmrit/logixng/implementation/DefaultFemaleGenericExpressionSocket.java
@@ -11,7 +11,6 @@ import javax.annotation.CheckForNull;
 import jmri.InstanceManager;
 import jmri.JmriException;
 import jmri.jmrit.logixng.*;
-import jmri.util.TypeConversionUtil;
 
 /**
  * Default implementation of the FemaleGenericExpressionSocket
@@ -22,9 +21,7 @@ public class DefaultFemaleGenericExpressionSocket
 
     private SocketType _socketType;             // The type of the socket the user has selected
     private SocketType _currentSocketType;      // The current type of the socket.
-    private FemaleSocket _internalSocket;       // The socket that the owner of this female socket uses.
     private FemaleSocket _currentActiveSocket;  // The socket that is currently in use, if any. Null otherwise.
-    private final FemaleSocketListener _socketListener;
     private final FemaleAnalogExpressionSocket _analogSocket = new DefaultFemaleAnalogExpressionSocket(this, this, "A");
     private final FemaleDigitalExpressionSocket _digitalSocket = new DefaultFemaleDigitalExpressionSocket(this, this, "D");
     private final FemaleStringExpressionSocket _stringSocket = new DefaultFemaleStringExpressionSocket(this, this, "S");
@@ -40,7 +37,6 @@ public class DefaultFemaleGenericExpressionSocket
         
         _socketType = socketType;
         _currentSocketType = socketType;
-        _socketListener = listener;
         
         switch (_socketType) {
             case ANALOG:
@@ -71,44 +67,6 @@ public class DefaultFemaleGenericExpressionSocket
         return _currentActiveSocket;
     }
     
-    
-    public FemaleAnalogExpressionSocket getAnalogSocket(Base parent) {
-        if (_internalSocket != null) {
-            throw new RuntimeException("internal socket cannot be set more than once");
-        }
-        FemaleAnalogExpressionSocket socket = new AnalogSocket(parent, _socketListener, getName());
-        _internalSocket = socket;
-        return socket;
-    }
-    
-    public FemaleDigitalExpressionSocket getDigitalSocket(Base parent) {
-        if (_internalSocket != null) {
-            throw new RuntimeException("internal socket cannot be set more than once");
-        }
-        
-        FemaleDigitalExpressionSocket socket = new DigitalSocket(parent, _socketListener, getName());
-        _internalSocket = socket;
-        return socket;
-    }
-    
-    public FemaleStringExpressionSocket getStringSocket(Base parent) {
-        if (_internalSocket != null) {
-            throw new RuntimeException("internal socket cannot be set more than once");
-        }
-        FemaleStringExpressionSocket socket = new StringSocket(parent, _socketListener, getName());
-        _internalSocket = socket;
-        return socket;
-    }
-    
-    public FemaleGenericExpressionSocket getGenericSocket(Base parent) {
-        if (_internalSocket != null) {
-            throw new RuntimeException("internal socket cannot be set more than once");
-        }
-        
-        FemaleGenericExpressionSocket socket = new GenericSocket(parent, _socketListener, getName());
-        _internalSocket = socket;
-        return socket;
-    }
     
     /** {@inheritDoc} */
     @Override
@@ -175,82 +133,6 @@ public class DefaultFemaleGenericExpressionSocket
         return _do_i18n;
     }
     
-    /** {@inheritDoc} */
-    @Override
-    public boolean evaluateBoolean() throws JmriException {
-        if (isConnected()) {
-            switch (_currentSocketType) {
-                case DIGITAL:
-                    return ((MaleDigitalExpressionSocket)getConnectedSocket()).evaluate();
-
-                case ANALOG:
-                    return TypeConversionUtil.convertToBoolean(
-                            ((MaleAnalogExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                case STRING:
-                    return TypeConversionUtil.convertToBoolean(
-                            ((MaleStringExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                default:
-                    throw new RuntimeException("socketType has invalid value: "+_currentSocketType.name());
-            }
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public double evaluateDouble() throws JmriException {
-        if (isConnected()) {
-            switch (_currentSocketType) {
-                case DIGITAL:
-                    return TypeConversionUtil.convertToDouble(
-                            ((MaleDigitalExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                case ANALOG:
-                    return ((MaleAnalogExpressionSocket)getConnectedSocket()).evaluate();
-
-                case STRING:
-                    return TypeConversionUtil.convertToDouble(
-                            ((MaleStringExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                default:
-                    throw new RuntimeException("socketType has invalid value: "+_currentSocketType.name());
-            }
-        } else {
-            return 0.0f;
-        }
-    }
-
-    @Override
-    public String evaluateString() throws JmriException {
-        if (isConnected()) {
-            switch (_currentSocketType) {
-                case DIGITAL:
-                    return TypeConversionUtil.convertToString(
-                            ((MaleDigitalExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                case ANALOG:
-                    return TypeConversionUtil.convertToString(
-                            ((MaleAnalogExpressionSocket)getConnectedSocket()).evaluate(),
-                            _do_i18n);
-
-                case STRING:
-                    return ((MaleStringExpressionSocket)getConnectedSocket()).evaluate();
-
-                default:
-                    throw new RuntimeException("socketType has invalid value: "+_currentSocketType.name());
-            }
-        } else {
-            return "";
-        }
-    }
-
     @Override
     @CheckForNull
     public Object evaluateGeneric() throws JmriException {
@@ -332,6 +214,7 @@ public class DefaultFemaleGenericExpressionSocket
                 throw new SocketAlreadyConnectedException("Socket is already connected");
             } else {
                 _currentActiveSocket.connect(socket);
+                _listener.connected(this);
                 return;
             }
         }
@@ -352,6 +235,7 @@ public class DefaultFemaleGenericExpressionSocket
             throw new IllegalArgumentException("Socket is not compatible");
         }
         _currentActiveSocket.connect(socket);
+        _listener.connected(this);
     }
 
     /** {@inheritDoc} */
@@ -361,6 +245,7 @@ public class DefaultFemaleGenericExpressionSocket
                 && _currentActiveSocket.isConnected()) {
             
             _currentActiveSocket.disconnect();
+            _listener.disconnected(this);
         }
     }
     
@@ -392,568 +277,10 @@ public class DefaultFemaleGenericExpressionSocket
         }
     }
     
-    /*.* {@inheritDoc} *./
-    @Override
-    protected void registerListenersForThisClass() {
-        if (_currentActiveSocket != null) {
-            _currentActiveSocket.registerListeners();
-        }
-    }
-    
-    /*.* {@inheritDoc} *./
-    @Override
-    protected void unregisterListenersForThisClass() {
-        if (_currentActiveSocket != null) {
-            _currentActiveSocket.unregisterListeners();
-        }
-    }
-*/
     /** {@inheritDoc} */
     @Override
     public void disposeMe() {
-//        if ((_currentActiveSocket != null) && (_currentActiveSocket.isConnected())) {
-//            MaleSocket aSocket = _currentActiveSocket.getConnectedSocket();
-//            disconnect();
-//            aSocket.dispose();
-//        }
+        // Do nothing
     }
     
-    
-    
-    private class AnalogSocket extends DefaultFemaleAnalogExpressionSocket {
-        
-        public AnalogSocket(Base parent, FemaleSocketListener listener, String name) {
-            super(parent, listener, name);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean isCompatible(MaleSocket socket) {
-            return DefaultFemaleGenericExpressionSocket.this.isCompatible(socket);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public double evaluate() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateDouble();
-        }
-
-        @Override
-        public Map<Category, List<Class<? extends Base>>> getConnectableClasses() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectableClasses();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Base getParent() {
-            return DefaultFemaleGenericExpressionSocket.this.getParent();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setParent(Base parent) {
-            DefaultFemaleGenericExpressionSocket.this.setParent(parent);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Lock getLock() {
-            return DefaultFemaleGenericExpressionSocket.this.getLock();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setLock(Lock lock) {
-            DefaultFemaleGenericExpressionSocket.this.setLock(lock);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void connect(MaleSocket socket) throws SocketAlreadyConnectedException {
-            DefaultFemaleGenericExpressionSocket.this.connect(socket);
-            _listener.connected(this);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disconnect() {
-            if (DefaultFemaleGenericExpressionSocket.this.isConnected()) {
-                DefaultFemaleGenericExpressionSocket.this.disconnect();
-                _listener.disconnected(this);
-            }
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public MaleSocket getConnectedSocket() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectedSocket();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public boolean isConnected() {
-            return DefaultFemaleGenericExpressionSocket.this.isConnected();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setName(String name) {
-            DefaultFemaleGenericExpressionSocket.this.setName(name);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getName() {
-            return DefaultFemaleGenericExpressionSocket.this.getName();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disposeMe() {
-            DefaultFemaleGenericExpressionSocket.this.dispose();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void registerListeners() {
-            DefaultFemaleGenericExpressionSocket.this.registerListeners();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void unregisterListeners() {
-            DefaultFemaleGenericExpressionSocket.this.unregisterListeners();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getShortDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getShortDescription(locale);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getLongDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getLongDescription(locale);
-        }
-
-    }
-    
-    
-    private class DigitalSocket extends DefaultFemaleDigitalExpressionSocket {
-        
-        public DigitalSocket(Base parent, FemaleSocketListener listener, String name) {
-            super(parent, listener, name);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean isCompatible(MaleSocket socket) {
-            return DefaultFemaleGenericExpressionSocket.this.isCompatible(socket);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean evaluate() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateBoolean();
-        }
-
-        @Override
-        public Map<Category, List<Class<? extends Base>>> getConnectableClasses() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectableClasses();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Base getParent() {
-            return DefaultFemaleGenericExpressionSocket.this.getParent();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setParent(Base parent) {
-            DefaultFemaleGenericExpressionSocket.this.setParent(parent);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Lock getLock() {
-            return DefaultFemaleGenericExpressionSocket.this.getLock();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setLock(Lock lock) {
-            DefaultFemaleGenericExpressionSocket.this.setLock(lock);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void connect(MaleSocket socket) throws SocketAlreadyConnectedException {
-            DefaultFemaleGenericExpressionSocket.this.connect(socket);
-            _listener.connected(this);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disconnect() {
-            if (DefaultFemaleGenericExpressionSocket.this.isConnected()) {
-                DefaultFemaleGenericExpressionSocket.this.disconnect();
-                _listener.disconnected(this);
-            }
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public MaleSocket getConnectedSocket() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectedSocket();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public boolean isConnected() {
-            return DefaultFemaleGenericExpressionSocket.this.isConnected();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setName(String name) {
-            DefaultFemaleGenericExpressionSocket.this.setName(name);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getName() {
-            return DefaultFemaleGenericExpressionSocket.this.getName();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disposeMe() {
-            DefaultFemaleGenericExpressionSocket.this.dispose();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void registerListeners() {
-            DefaultFemaleGenericExpressionSocket.this.registerListeners();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void unregisterListeners() {
-            DefaultFemaleGenericExpressionSocket.this.unregisterListeners();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getShortDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getShortDescription(locale);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getLongDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getLongDescription(locale);
-        }
-
-    }
-    
-    
-    private class StringSocket extends DefaultFemaleStringExpressionSocket {
-        
-        public StringSocket(Base parent, FemaleSocketListener listener, String name) {
-            super(parent, listener, name);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean isCompatible(MaleSocket socket) {
-            return DefaultFemaleGenericExpressionSocket.this.isCompatible(socket);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public String evaluate() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateString();
-        }
-
-        @Override
-        public Map<Category, List<Class<? extends Base>>> getConnectableClasses() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectableClasses();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Base getParent() {
-            return DefaultFemaleGenericExpressionSocket.this.getParent();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setParent(Base parent) {
-            DefaultFemaleGenericExpressionSocket.this.setParent(parent);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Lock getLock() {
-            return DefaultFemaleGenericExpressionSocket.this.getLock();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setLock(Lock lock) {
-            DefaultFemaleGenericExpressionSocket.this.setLock(lock);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void connect(MaleSocket socket) throws SocketAlreadyConnectedException {
-            DefaultFemaleGenericExpressionSocket.this.connect(socket);
-            _listener.connected(this);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disconnect() {
-            if (DefaultFemaleGenericExpressionSocket.this.isConnected()) {
-                DefaultFemaleGenericExpressionSocket.this.disconnect();
-                _listener.disconnected(this);
-            }
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public MaleSocket getConnectedSocket() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectedSocket();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public boolean isConnected() {
-            return DefaultFemaleGenericExpressionSocket.this.isConnected();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setName(String name) {
-            DefaultFemaleGenericExpressionSocket.this.setName(name);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getName() {
-            return DefaultFemaleGenericExpressionSocket.this.getName();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disposeMe() {
-            DefaultFemaleGenericExpressionSocket.this.dispose();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void registerListeners() {
-            DefaultFemaleGenericExpressionSocket.this.registerListeners();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void unregisterListeners() {
-            DefaultFemaleGenericExpressionSocket.this.unregisterListeners();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getShortDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getShortDescription(locale);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getLongDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getLongDescription(locale);
-        }
-
-    }
-    
-    
-    private class GenericSocket extends AbstractFemaleSocket
-            implements FemaleGenericExpressionSocket {
-        
-        public GenericSocket(Base parent, FemaleSocketListener listener, String name) {
-            super(parent, listener, name);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public FemaleSocket getCurrentActiveSocket() {
-            return DefaultFemaleGenericExpressionSocket.this.getCurrentActiveSocket();
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public void setSocketType(SocketType socketType)
-                throws SocketAlreadyConnectedException {
-            DefaultFemaleGenericExpressionSocket.this.setSocketType(socketType);
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public SocketType getSocketType() {
-            return DefaultFemaleGenericExpressionSocket.this.getSocketType();
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean isCompatible(MaleSocket socket) {
-            return DefaultFemaleGenericExpressionSocket.this.isCompatible(socket);
-        }
-        
-        
-        /** {@inheritDoc} */
-        @Override
-        public boolean evaluateBoolean() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateBoolean();
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public double evaluateDouble() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateDouble();
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public String evaluateString() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateString();
-        }
-        
-        /** {@inheritDoc} */
-        @Override
-        public Object evaluateGeneric() throws JmriException {
-            return DefaultFemaleGenericExpressionSocket.this.evaluateGeneric();
-        }
-
-        @Override
-        public Map<Category, List<Class<? extends Base>>> getConnectableClasses() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectableClasses();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Base getParent() {
-            return DefaultFemaleGenericExpressionSocket.this.getParent();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setParent(Base parent) {
-            DefaultFemaleGenericExpressionSocket.this.setParent(parent);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public Lock getLock() {
-            return DefaultFemaleGenericExpressionSocket.this.getLock();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setLock(Lock lock) {
-            DefaultFemaleGenericExpressionSocket.this.setLock(lock);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void connect(MaleSocket socket) throws SocketAlreadyConnectedException {
-            DefaultFemaleGenericExpressionSocket.this.connect(socket);
-            _listener.connected(this);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disconnect() {
-            if (DefaultFemaleGenericExpressionSocket.this.isConnected()) {
-                DefaultFemaleGenericExpressionSocket.this.disconnect();
-                _listener.disconnected(this);
-            }
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public MaleSocket getConnectedSocket() {
-            return DefaultFemaleGenericExpressionSocket.this.getConnectedSocket();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public boolean isConnected() {
-            return DefaultFemaleGenericExpressionSocket.this.isConnected();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void setName(String name) {
-            DefaultFemaleGenericExpressionSocket.this.setName(name);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getName() {
-            return DefaultFemaleGenericExpressionSocket.this.getName();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public void disposeMe() {
-            DefaultFemaleGenericExpressionSocket.this.dispose();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void registerListeners() {
-            DefaultFemaleGenericExpressionSocket.this.registerListeners();
-        }
-
-        /**
-         * Register listeners if this object needs that.
-         */
-        @Override
-        public void unregisterListeners() {
-            DefaultFemaleGenericExpressionSocket.this.unregisterListeners();
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getShortDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getShortDescription(locale);
-        }
-
-        /** {@inheritDoc} */
-        @Override
-        public String getLongDescription(Locale locale) {
-            return DefaultFemaleGenericExpressionSocket.this.getLongDescription(locale);
-        }
-
-    }
-    
-//    private final static org.slf4j.Logger log =
-//            org.slf4j.LoggerFactory.getLogger(DefaultFemaleGenericExpressionSocket.class);
 }
