+++ java/test/jmri/jmrit/logixng/expressions/DigitalFormulaTest.java
@@ -177,7 +177,7 @@ public class DigitalFormulaTest extends AbstractDigitalExpressionTestBase {
             Assert.assertEquals("expression female socket name is "+socketData._socketName,
                     socketData._socketName, expression.getChild(i).getName());
             Assert.assertEquals("expression female socket is of correct class",
-                    "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket$GenericSocket",
+                    "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket",
                     expression.getChild(i).getClass().getName());
             Assert.assertFalse("expression female socket is not connected",
                     expression.getChild(i).isConnected());
@@ -236,7 +236,7 @@ public class DigitalFormulaTest extends AbstractDigitalExpressionTestBase {
             Assert.assertEquals("expression female socket name is "+socketData._socketName,
                     socketData._socketName, expression.getChild(i).getName());
             Assert.assertEquals("expression female socket is of correct class",
-                    "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket$GenericSocket",
+                    "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket",
                     expression.getChild(i).getClass().getName());
             Assert.assertFalse("expression female socket is not connected",
                     expression.getChild(i).isConnected());
