+++ java/test/jmri/jmrit/logixng/implementation/DefaultFemaleGenericExpressionSocket2_Test.java
@@ -24,9 +24,8 @@ public class DefaultFemaleGenericExpressionSocket2_Test {
     private FemaleSocketListener _listener;
     
     @Test
-    public void testEvaluateBoolean() throws JmriException {
+    public void testEvaluateGeneric() throws JmriException {
         DefaultFemaleGenericExpressionSocket socket;
-        FemaleGenericExpressionSocket internalGenericSocket;
         
         MyAnalogExpression analogExpression = new MyAnalogExpression("IQAE351", null);
         MaleSocket analogMaleSocket =
@@ -40,199 +39,43 @@ public class DefaultFemaleGenericExpressionSocket2_Test {
         MaleSocket stringMaleSocket =
                 InstanceManager.getDefault(StringExpressionManager.class).registerExpression(stringExpression);
         
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.ANALOG, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
+        socket = new DefaultFemaleGenericExpressionSocket(SocketType.GENERIC, _conditionalNG, _listener, "E");
         
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertEquals("evaluateGeneric() returns correct value", null, socket.evaluateGeneric());
         
         socket.connect(analogMaleSocket);
         analogExpression._value = 0.0;
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertTrue("evaluateGeneric() returns correct value", 0.0 == (Double)socket.evaluateGeneric());
         analogExpression._value = 1.0;
-        Assert.assertTrue("evaluateBoolean() returns true", socket.evaluateBoolean());
-        Assert.assertTrue("evaluateBoolean() returns true", internalGenericSocket.evaluateBoolean());
+        Assert.assertTrue("evaluateGeneric() returns correct value", 1.0 == (Double)socket.evaluateGeneric());
         analogExpression._value = -1.0;
-        Assert.assertTrue("evaluateBoolean() returns true", socket.evaluateBoolean());
-        Assert.assertTrue("evaluateBoolean() returns true", internalGenericSocket.evaluateBoolean());
+        Assert.assertTrue("evaluateGeneric() returns correct value", -1.0 == (Double)socket.evaluateGeneric());
         socket.disconnect();
         
         
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.DIGITAL, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
+        socket = new DefaultFemaleGenericExpressionSocket(SocketType.GENERIC, _conditionalNG, _listener, "E");
         
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertEquals("evaluateGeneric() returns correct value", null, socket.evaluateGeneric());
         
         socket.connect(digitalMaleSocket);
         digitalExpression._value = false;
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertFalse("evaluateGeneric() returns correct value", (Boolean)socket.evaluateGeneric());
         digitalExpression._value = true;
-        Assert.assertTrue("evaluateBoolean() returns true", socket.evaluateBoolean());
-        Assert.assertTrue("evaluateBoolean() returns true", internalGenericSocket.evaluateBoolean());
+        Assert.assertTrue("evaluateGeneric() returns correct value", (Boolean)socket.evaluateGeneric());
         socket.disconnect();
         
         
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.STRING, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
+        socket = new DefaultFemaleGenericExpressionSocket(SocketType.GENERIC, _conditionalNG, _listener, "E");
         
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertEquals("evaluateGeneric() returns correct value", null, socket.evaluateGeneric());
         
         socket.connect(stringMaleSocket);
         stringExpression._value = "";
-        Assert.assertFalse("evaluateBoolean() returns false", socket.evaluateBoolean());
-        Assert.assertFalse("evaluateBoolean() returns false", internalGenericSocket.evaluateBoolean());
+        Assert.assertEquals("evaluateGeneric() returns correct value", "", socket.evaluateGeneric());
         stringExpression._value = "Hello";
-        Assert.assertTrue("evaluateBoolean() returns true", socket.evaluateBoolean());
-        Assert.assertTrue("evaluateBoolean() returns true", internalGenericSocket.evaluateBoolean());
+        Assert.assertEquals("evaluateGeneric() returns correct value", "Hello", socket.evaluateGeneric());
         stringExpression._value = "1.0";
-        Assert.assertTrue("evaluateBoolean() returns true", socket.evaluateBoolean());
-        Assert.assertTrue("evaluateBoolean() returns true", internalGenericSocket.evaluateBoolean());
-        socket.disconnect();
-    }
-    
-    @Test
-    public void testEvaluateDouble() throws JmriException {
-        DefaultFemaleGenericExpressionSocket socket;
-        FemaleGenericExpressionSocket internalGenericSocket;
-        
-        MyAnalogExpression analogExpression = new MyAnalogExpression("IQAE351", null);
-        MaleSocket analogMaleSocket =
-                InstanceManager.getDefault(AnalogExpressionManager.class).registerExpression(analogExpression);
-        
-        MyDigitalExpression digitalExpression = new MyDigitalExpression("IQDE351", null);
-        MaleSocket digitalMaleSocket =
-                InstanceManager.getDefault(DigitalExpressionManager.class).registerExpression(digitalExpression);
-        
-        MyStringExpression stringExpression = new MyStringExpression("IQSE351", null);
-        MaleSocket stringMaleSocket =
-                InstanceManager.getDefault(StringExpressionManager.class).registerExpression(stringExpression);
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.ANALOG, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertTrue("evaluateBoolean() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateBoolean() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        
-        socket.connect(analogMaleSocket);
-        analogExpression._value = 0.0;
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        analogExpression._value = 1.0;
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == internalGenericSocket.evaluateDouble());
-        analogExpression._value = -1.0;
-        Assert.assertTrue("evaluateDouble() returns -1.0", -1.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns -1.0", -1.0 == internalGenericSocket.evaluateDouble());
-        socket.disconnect();
-        
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.DIGITAL, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        
-        socket.connect(digitalMaleSocket);
-        digitalExpression._value = false;
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        digitalExpression._value = true;
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == internalGenericSocket.evaluateDouble());
-        socket.disconnect();
-        
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.STRING, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        
-        socket.connect(stringMaleSocket);
-        stringExpression._value = "";
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        stringExpression._value = "Hello";
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 0.0", 0.0 == internalGenericSocket.evaluateDouble());
-        stringExpression._value = "1.0";
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == socket.evaluateDouble());
-        Assert.assertTrue("evaluateDouble() returns 1.0", 1.0 == internalGenericSocket.evaluateDouble());
-        socket.disconnect();
-    }
-    
-    @Test
-    public void testEvaluateString() throws JmriException {
-        DefaultFemaleGenericExpressionSocket socket;
-        FemaleGenericExpressionSocket internalGenericSocket;
-        
-        MyAnalogExpression analogExpression = new MyAnalogExpression("IQAE351", null);
-        MaleSocket analogMaleSocket =
-                InstanceManager.getDefault(AnalogExpressionManager.class).registerExpression(analogExpression);
-        
-        MyDigitalExpression digitalExpression = new MyDigitalExpression("IQDE351", null);
-        MaleSocket digitalMaleSocket =
-                InstanceManager.getDefault(DigitalExpressionManager.class).registerExpression(digitalExpression);
-        
-        MyStringExpression stringExpression = new MyStringExpression("IQSE351", null);
-        MaleSocket stringMaleSocket =
-                InstanceManager.getDefault(StringExpressionManager.class).registerExpression(stringExpression);
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.ANALOG, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertEquals("evaluateString() returns empty string", "", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns empty string", "", internalGenericSocket.evaluateString());
-        
-        socket.connect(analogMaleSocket);
-        analogExpression._value = 0.0;
-        Assert.assertEquals("evaluateString() returns 0.0", "0.0", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns 0.0", "0.0", internalGenericSocket.evaluateString());
-        analogExpression._value = 1.0;
-        Assert.assertEquals("evaluateString() returns 1.0", "1.0", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns 1.0", "1.0", internalGenericSocket.evaluateString());
-        analogExpression._value = -1.0;
-        Assert.assertEquals("evaluateString() returns -1.0", "-1.0", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns -1.0", "-1.0", internalGenericSocket.evaluateString());
-        socket.disconnect();
-        
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.DIGITAL, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertEquals("evaluateString() returns empty string", "", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns empty string", "", internalGenericSocket.evaluateString());
-        
-        socket.connect(digitalMaleSocket);
-        digitalExpression._value = false;
-        Assert.assertEquals("evaluateString() returns false", "false", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns false", "false", internalGenericSocket.evaluateString());
-        digitalExpression._value = true;
-        Assert.assertEquals("evaluateString() returns true", "true", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns true", "true", internalGenericSocket.evaluateString());
-        socket.disconnect();
-        
-        
-        socket = new DefaultFemaleGenericExpressionSocket(SocketType.STRING, _conditionalNG, _listener, "E");
-        internalGenericSocket = socket.getGenericSocket(_conditionalNG);
-        
-        Assert.assertEquals("evaluateString() returns empty string", "", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns empty string", "", internalGenericSocket.evaluateString());
-        
-        socket.connect(stringMaleSocket);
-        stringExpression._value = "";
-        Assert.assertEquals("evaluateString() returns empty string", "", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns empty string", "", internalGenericSocket.evaluateString());
-        stringExpression._value = "Hello";
-        Assert.assertEquals("evaluateString() returns Hello", "Hello", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns Hello", "Hello", internalGenericSocket.evaluateString());
-        stringExpression._value = "1.0";
-        Assert.assertEquals("evaluateString() returns 1.0", "1.0", socket.evaluateString());
-        Assert.assertEquals("evaluateString() returns 1.0", "1.0", internalGenericSocket.evaluateString());
+        Assert.assertEquals("evaluateGeneric() returns correct value", "1.0", socket.evaluateGeneric());
         socket.disconnect();
     }
