+++ java/test/jmri/jmrit/logixng/implementation/DigitalExpressionManagerTest.java
@@ -58,15 +58,9 @@ public class DigitalExpressionManagerTest extends AbstractManagerTestBase {
         FemaleSocket socket;
         MyExpression myExpression = new MyExpression("IQSA1");
         FemaleSocketListener listener = new MyFemaleSocketListener();
-        LogixNGPreferences preferences = InstanceManager.getDefault(LogixNGPreferences.class);
         
-        preferences.setUseGenericFemaleSockets(false);
         socket = _m.createFemaleSocket(myExpression, listener, "E1");
         Assert.assertEquals("Class is correct", "jmri.jmrit.logixng.implementation.DefaultFemaleDigitalExpressionSocket", socket.getClass().getName());
-        
-        preferences.setUseGenericFemaleSockets(true);
-        socket = _m.createFemaleSocket(myExpression, listener, "E2");
-        Assert.assertEquals("Class is correct", "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket$DigitalSocket", socket.getClass().getName());
     }
     
     @Test
