+++ java/test/jmri/jmrit/logixng/implementation/StringExpressionManagerTest.java
@@ -60,13 +60,8 @@ public class StringExpressionManagerTest extends AbstractManagerTestBase {
         FemaleSocketListener listener = new MyFemaleSocketListener();
         LogixNGPreferences preferences = InstanceManager.getDefault(LogixNGPreferences.class);
         
-        preferences.setUseGenericFemaleSockets(false);
         socket = _m.createFemaleSocket(myExpression, listener, "E1");
         Assert.assertEquals("Class is correct", "jmri.jmrit.logixng.implementation.DefaultFemaleStringExpressionSocket", socket.getClass().getName());
-        
-        preferences.setUseGenericFemaleSockets(true);
-        socket = _m.createFemaleSocket(myExpression, listener, "E2");
-        Assert.assertEquals("Class is correct", "jmri.jmrit.logixng.implementation.DefaultFemaleGenericExpressionSocket$StringSocket", socket.getClass().getName());
     }
     
     @Test
